# api/openapi.yaml
# API contract — owned by THIS repo (contract-per-repo). Phase 2a: real domain.
openapi: 3.0.3
info:
  title: "Current Account — BIAN Semantic API"
  description: >
    BIAN Service Domain 'Current Account' (Fulfill) operating on the
    'Current Account Facility Fulfillment Arrangement' control record.
    Deep implementation: deposits/withdrawals with overdraft, KYC gating,
    block/close lifecycle. All money is integer minor units.
  version: 0.2.0
servers:
  - url: /
  - url: /sd-current-account
    description: through the platform gateway

paths:
  /v1/service-domain:
    get:
      operationId: getServiceDomain
      summary: Service domain metadata
      responses:
        "200": { description: SD identity and BIAN classification }

  /v1/current-account-facility-fulfillment-arrangement/initiate:
    post:
      operationId: openAccount
      summary: "Initiate — open a current account (starts PENDING_KYC)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerReference]
              properties:
                customerReference: { type: string }
                currency: { type: string, pattern: "^[A-Z]{3}$", default: INR }
                overdraftLimitMinor: { type: integer, format: int64, minimum: 0, default: 0 }
      responses:
        "201":
          description: Account opened (ACTIVE if KYC auto-approve is on, else PENDING_KYC)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/Invalid" }

  /v1/current-account-facility-fulfillment-arrangement:
    get:
      operationId: listAccounts
      summary: Retrieve (collection)
      responses:
        "200":
          description: All accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Account" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/retrieve:
    get:
      operationId: retrieveAccount
      summary: Retrieve
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "404": { $ref: "#/components/responses/NotFound" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/control:
    put:
      operationId: controlAccount
      summary: "Control — block | unblock | terminate (close; requires balance 0)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [block, unblock, terminate] }
      responses:
        "200": { description: New account state }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/kyc-result:
    put:
      operationId: applyKycResult
      summary: "KYC outcome callback (from Know Your Customer SD; Kafka consumer later)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved: { type: boolean }
                reason: { type: string }
      responses:
        "200": { description: ACTIVE on approval, REJECTED (terminal) on rejection }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/payments/deposit:
    post:
      operationId: deposit
      summary: "Payments BQ — credit funds"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody: { $ref: "#/components/requestBodies/Posting" }
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/payments/withdraw:
    post:
      operationId: withdraw
      summary: "Payments BQ — debit funds (may draw into overdraft, never beyond)"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody: { $ref: "#/components/requestBodies/Posting" }
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/payments/cheque-credit:
    post:
      operationId: chequeCredit
      summary: "Inbound credit from Cheque Processing on cheque.cleared"
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      requestBody: { $ref: "#/components/requestBodies/Posting" }
      responses:
        "201": { $ref: "#/components/responses/Posted" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/payments:
    get:
      operationId: listTransactions
      summary: Transaction history (running balance per posting)
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200":
          description: Postings, oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }

  /v1/current-account-facility-fulfillment-arrangement/{accountId}/balance:
    get:
      operationId: getBalance
      summary: Balance + available (balance plus undrawn overdraft)
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200": { description: Balance view }
        "404": { $ref: "#/components/responses/NotFound" }

components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string }

  requestBodies:
    Posting:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [amountMinor]
            properties:
              amountMinor: { type: integer, format: int64, minimum: 1 }
              reference: { type: string }

  responses:
    Posted:
      description: Transaction posted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Transaction" }
    NotFound:
      description: Unknown account
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RuleViolation:
      description: Business rule forbids this (KYC pending, blocked, overdraft, non-zero close…)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: Bad input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  schemas:
    Account:
      type: object
      properties:
        accountId: { type: string, example: "CA-…" }
        customerReference: { type: string }
        currency: { type: string, example: INR }
        balanceMinor: { type: integer, format: int64 }
        overdraftLimitMinor: { type: integer, format: int64 }
        availableMinor: { type: integer, format: int64 }
        status:
          type: string
          enum: [PENDING_KYC, ACTIVE, BLOCKED, REJECTED, CLOSED]
        openedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Transaction:
      type: object
      properties:
        transactionId: { type: string }
        accountId: { type: string }
        type: { type: string, enum: [DEPOSIT, WITHDRAWAL, FEE, CHEQUE_CREDIT] }
        amountMinor: { type: integer, format: int64, description: "signed: credits +, debits -" }
        balanceAfterMinor: { type: integer, format: int64 }
        reference: { type: string }
        postedAt: { type: string, format: date-time }

    Error:
      type: object
      properties:
        code: { type: string, example: OVERDRAFT_EXCEEDED }
        message: { type: string }